Controller, validationÂ and route - Transactions
