+++ routes/index.js
@@ -6,6 +6,6 @@ const router = express.Router()
 
 // example of a route with index controller get function
 // router.get('/', get)
-router.use('/category', categoryRouter)
+router.use('/', categoryRouter)
 
 module.exports = router
